[NavTabCarousel] Fix component tree view + minor style issue (#1519)
