+++ assets/javascripts/kitten/components/carousel/nav-tab-carousel/components/next.js
