+++ assets/javascripts/kitten/components/carousel/nav-tab-carousel/components/prev.js
